+++ packages/onboarding/src/steps/ComposableScreen/elements/InputElement.ts
@@ -13,6 +13,7 @@ export type InputElementProps = BaseBoxProps & {
   multiline?: boolean;
   numberOfLines?: number;
   editable?: boolean;
+  autoFocus?: boolean;
   color?: string;
   backgroundColor?: string;
   fontSize?: number;
@@ -41,6 +42,7 @@ export const InputElementPropsSchema = BaseBoxPropsSchema.extend({
   multiline: z.boolean().optional(),
   numberOfLines: z.number().int().nonnegative().optional(),
   editable: z.boolean().optional(),
+  autoFocus: z.boolean().optional(),
   color: z.string().optional(),
   backgroundColor: z.string().optional(),
   fontSize: z.number().optional(),